kis_mcp_server/requirements.txt
fastmcp
requests
python-dotenv
pandas
